drm: imx8: fix hdmi firmware load
Loading the hdmi firmware from kernel is broken. The firmware is loaded
from filesystem but never sent to the hdmi controller. This commit fixes
it by waiting for the firmware load to complete and afterwards transfer
it to the controller.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>

Upstream-Status: Pending

Downstream specific, i.MX 8 HDMI subsystem not yet in upstream.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>