tty: serial: fsl_lpuart: set rxiden in non dma case
In the case of not using edma set the RXIDEN field. This sets the RDRF
flag even if there are received characters in the FIFO and the receiver
has been idle for the specified number of characters.

Related-to: ELB-3517
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
(cherry picked from commit 35a34002c3c946195d72a12aa7cae91a21cdfb0a)
Upstream-Status: Inappropriate [Downstream specific, it requires additional patches on fsl_lpuart not available in upstream]
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>