sec_mipi_dsim-imx: properly clean up if probe fails
It is possible that component_add() does return -EPROBE_DEFER if
something else is not ready. At that stage of the probe
sec_dsim_of_parse_resets() has already been called so the resets need to
be cleaned up to leave it in a clean state for the next try.

Upstream-Status: Pending [downstream graphics stack]
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>