toradex_imx_v8.config: make apalis imx8 hdmi driver a module
This is needed so that the HDMI FW loader has a mounted rootfs as
we do not load the FW in U-Boot.

Additonally remove the HDMI receiver driver. It does not compile
if the HDMI transmitter is compiled as a module and the HW pins
needed are not accessible on a Apalis iMX8.

Upstream-Status: Inappropriate [Configuration]
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>