drm/imx: lcdifv3: crtc increase pixel clock tolerance to 10%
Set a pixel clock tolerance of 10% for the DSI interface of the i.MX 8MP
SoCs. There are some resolutions e.g. 1280x1024@60 Hz that do not work
without this commit.

Upstream-Status: Pending

This driver is not upstream and probably never will. It looks like the
function will become part of mxsfb. They don't verify the pixel clock
frequency there.

Related-to: ELB-4728
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>